backend/requirements.txt
fastapi==0.115.0
uvicorn==0.34.0
apify-client==1.8.1
supabase==2.13.0
apscheduler==3.10.4
python-telegram-bot==21.10
httpx==0.28.1
numpy==2.2.4
python-dateutil==2.9.0
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.25.3
